Changed update of devlists/* to be performed only during full OSS build

+++ setup/FreeBSD/build.sh
@@ -55,9 +55,11 @@ cp lib/libOSSlib/libOSSlib.so prototype/usr/lib/oss/lib
 
 cp devlist.txt prototype/usr/lib/oss/etc/devices.list
 
-rm -f $SRCDIR/devlists/FreeBSD
-
-cp devlist.txt $SRCDIR/devlists/FreeBSD
+if test -d $SRCDIR/kernel/nonfree
+then
+	rm -f $SRCDIR/devlists/FreeBSD
+	cp devlist.txt $SRCDIR/devlists/FreeBSD
+fi
 
 # Generate Man pages for commands
 for i in target/bin/*
